+++ src/onegov.core/onegov/core/utils.py
@@ -4,6 +4,7 @@ import inspect
 import json
 import magic
 import mimetypes
+import morepath
 import os.path
 import pydoc
 import re
@@ -13,6 +14,7 @@ from contextlib import contextmanager
 from cProfile import Profile
 from datetime import datetime
 from functools import partial
+from importlib import import_module
 from itertools import groupby, tee
 from purl import URL
 from unidecode import unidecode
@@ -400,3 +402,37 @@ def is_sorted(iterable, key=lambda i: i, reverse=False):
             return False
 
     return True
+
+
+def morepath_modules(cls):
+    """ Returns all morepath modules which should be scanned for the given
+    morepath application class.
+
+    We can't reliably know the actual morepath modules that
+    need to be scanned, which is why we assume that each module has
+    one namespace (like 'more.transaction' or 'onegov.core').
+
+    """
+    for base in cls.__mro__:
+        if not issubclass(base, morepath.App):
+            continue
+
+        if base is morepath.App:
+            continue
+
+        module = '.'.join(base.__module__.split('.')[:2])
+
+        if module.startswith('test'):
+            continue
+
+        yield module
+
+
+def scan_morepath_modules(cls, config):
+    """ Tries to scann all the morepath modules required for the given
+    application class. This is not guaranteed to stay reliable as there is
+    no sure way to discover all modules required by the application class.
+
+    """
+    for module in sorted(morepath_modules(cls)):
+        config.scan(import_module(module))